Timeline Switch Component
New Scripts:
EnvironmentController: Controls the fade of the environment overlay and which environments are active(Past, present, future)
Tag Helper: Store all your tags here to avoid typing hardcoded strings.
Test_Anim: A temporary script while we are waiting on animations. The code in this will go in the player animation script.
Timeline: Simple enum Class
TimelineController: A component for UI that controls which timeline we are currently in. This controller manages the environment controller.